# openapi/orders.yaml
openapi: 3.0.3
info:
  title: Orders Service API
  version: 1.0.0
  description: >
    Orders service coordinates product and user information to register portfolio project orders.
servers:
  - url: http://localhost:8084/api
tags:
  - name: Orders
    description: Operations for managing purchase orders in the platform.
paths:
  /orders:
    get:
      tags: [Orders]
      summary: List orders
      operationId: listOrders
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
        - in: query
          name: size
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/OrderStatus'
      responses:
        '200':
          description: Orders page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
    post:
      tags: [Orders]
      summary: Create order
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Validation error
        '404':
          description: User or product not found
  /orders/{id}:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    get:
      tags: [Orders]
      summary: Get order by id
      operationId: getOrder
      responses:
        '200':
          description: Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Missing
    put:
      tags: [Orders]
      summary: Update order items
      operationId: updateOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Missing
    delete:
      tags: [Orders]
      summary: Cancel order
      operationId: cancelOrder
      responses:
        '204':
          description: Cancelled
  /orders/{id}/status:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    patch:
      tags: [Orders]
      summary: Update order status
      operationId: updateOrderStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderStatusRequest'
      responses:
        '200':
          description: Updated status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /orders/user/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Orders]
      summary: List orders by user
      operationId: listOrdersByUser
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrderId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Order:
      type: object
      required:
        - id
        - userId
        - status
        - totalAmount
        - currency
        - createdAt
        - items
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        userName:
          type: string
          description: Full name snapshot from users-service
        userEmail:
          type: string
          format: email
        status:
          $ref: '#/components/schemas/OrderStatus'
        totalAmount:
          type: number
          format: double
        currency:
          type: string
          example: EUR
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderItem:
      type: object
      required:
        - productId
        - quantity
        - price
      properties:
        productId:
          type: string
          format: uuid
        productName:
          type: string
          description: Product name snapshot from catalog-service
        quantity:
          type: integer
          minimum: 1
        price:
          type: number
          format: double
        title:
          type: string
    CreateOrderRequest:
      type: object
      required:
        - userId
        - currency
        - items
      properties:
        userId:
          type: string
          format: uuid
        currency:
          type: string
          minLength: 3
          maxLength: 3
        notes:
          type: string
          maxLength: 300
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateOrderItem'
    CreateOrderItem:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
    UpdateOrderRequest:
      type: object
      required:
        - notes
        - items
      properties:
        notes:
          type: string
          maxLength: 300
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateOrderItem'
    OrderStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'
    OrderStatus:
      type: string
      enum:
        - PENDING
        - CONFIRMED
        - SHIPPED
        - DELIVERED
        - CANCELLED
    OrderPage:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Order'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
security:
  - bearerAuth: []